fix: some optimisation
