+++ package.json
@@ -16,6 +16,7 @@
   },
   "dependencies": {
     "@quasar/extras": "^1.16.4",
+    "axios": "^1.8.4",
     "core-js": "^3.31.1",
     "npm": "^10.9.2",
     "quasar": "^2.16.0",
