+++ src/router/authentication.js
@@ -2,7 +2,15 @@ const authentication = [
   {
     path: '/login',
       component: () => import('pages/LoginComponent.vue'),
-  }
+  },
+  {
+    path: '/authentication/',
+    component: () => import('pages/LoginComponent.vue'),
+    children: [
+      { path: 'login', name: 'login', component: () => import('src/modules/security/LoginComponent.vue') },
+      { path: 'register', name: 'register', component: () => import('src/modules/security/LoginComponent.vue') },
+    ],
+  },
 ]
 
 export default authentication
